[WOOR-224] refactor: Subscriber에서 채널 이름 반환하도록 리팩토링
