+++ woorimap-notification/src/main/java/com/musseukpeople/woorimapnotification/notification/application/message/RedisPostSubscriber.java
@@ -3,20 +3,22 @@ package com.musseukpeople.woorimapnotification.notification.application.message;
 import java.io.IOException;
 
 import org.springframework.data.redis.connection.Message;
-import org.springframework.data.redis.connection.MessageListener;
+import org.springframework.data.redis.listener.ChannelTopic;
+import org.springframework.data.redis.listener.Topic;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.musseukpeople.woorimapnotification.notification.application.NotificationService;
-import com.musseukpeople.woorimapnotification.notification.domain.PostEvent;
+import com.musseukpeople.woorimapnotification.notification.domain.event.PostEvent;
 
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class RedisMessageSubscriber implements MessageListener {
+public class RedisPostSubscriber implements RedisSubscriber {
 
+    private static final ChannelTopic POST_CHANNEL = ChannelTopic.of("post");
     private static final String CONVERT_FAIL_MESSAGE = "메시지를 변환하지 못했습니다.";
 
     private final ObjectMapper objectMapper;
@@ -31,5 +33,10 @@ public class RedisMessageSubscriber implements MessageListener {
         } catch (IOException e) {
             throw new IllegalArgumentException(CONVERT_FAIL_MESSAGE, e);
         }
+    }
+
+    @Override
+    public Topic getTopic() {
+        return POST_CHANNEL;
     }
 }
