+++ woorimap-notification/src/main/java/com/musseukpeople/woorimapnotification/notification/application/message/RedisSubscriber.java
@@ -1,0 +1,9 @@
+package com.musseukpeople.woorimapnotification.notification.application.message;
+
+import org.springframework.data.redis.connection.MessageListener;
+import org.springframework.data.redis.listener.Topic;
+
+public interface RedisSubscriber extends MessageListener {
+
+    Topic getTopic();
+}
