+++ woorimap-notification/src/test/java/com/musseukpeople/woorimapnotification/notification/application/NotificationServiceTest.java
@@ -19,8 +19,8 @@ import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 import com.musseukpeople.woorimapnotification.notification.application.dto.response.NotificationResponse;
 import com.musseukpeople.woorimapnotification.notification.domain.Notification;
 import com.musseukpeople.woorimapnotification.notification.domain.NotificationRepository;
-import com.musseukpeople.woorimapnotification.notification.domain.PostEvent;
-import com.musseukpeople.woorimapnotification.notification.domain.PostEvent.EventType;
+import com.musseukpeople.woorimapnotification.notification.domain.event.PostEvent;
+import com.musseukpeople.woorimapnotification.notification.domain.event.PostEvent.EventType;
 import com.musseukpeople.woorimapnotification.notification.exception.NotFoundNotificationException;
 import com.musseukpeople.woorimapnotification.notification.infrastructure.EmitterRepository;
 import com.musseukpeople.woorimapnotification.notification.util.NotificationTest;
